% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggerror-package.R
\docType{package}
\name{ggerror-package}
\alias{ggerror}
\alias{ggerror-package}
\title{ggerror: Modified Error Geoms for 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Unifies the 'ggplot2' error geoms behind a single layer, geom_error(), that accepts an 'error' aesthetic, infers orientation from the data, computes min/max bounds, and dispatches to geom_errorbar(), geom_linerange(), geom_crossbar(), or geom_pointrange(). Adds support for asymmetric errors via 'error_neg' and 'error_pos' aesthetics, including one-sided error bars.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iamYannC/ggerror}
  \item Report bugs at \url{https://github.com/iamYannC/ggerror/issues}
}

}
\author{
\strong{Maintainer}: Yann Cohen \email{yannco5@gmail.com} (\href{https://orcid.org/0009-0009-0509-3609}{ORCID})

}
\keyword{internal}
