% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggerror-package.R
\docType{package}
\name{ggerror-package}
\alias{ggerror}
\alias{ggerror-package}
\title{ggerror: Ergonomic Error Bars and Range Indicators for ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a single geom layer, geom_error(), that wraps ggplot2 error bar geoms with automatic orientation inference. Users declare an error aesthetic and the package computes the min/max bounds and dispatches to geom_errorbar(), geom_linerange(), geom_crossbar(), or geom_pointrange().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iamYannC/ggerror}
  \item Report bugs at \url{https://github.com/iamYannC/ggerror/issues}
}

}
\author{
\strong{Maintainer}: Yann Cohen \email{yannco5@gmail.com} (\href{https://orcid.org/0009-0009-0509-3609}{ORCID})

}
\keyword{internal}
