% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggerror-package.R
\docType{package}
\name{ggerror-package}
\alias{ggerror}
\alias{ggerror-package}
\title{ggerror: Extended Error Geoms for 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extends the 'ggplot2' error geoms. geom_error() accepts an 'error' aesthetic with auto-inference of the orientation. It also supports 'error_neg' and 'error_pos' for asymmetric cases, with full control over aesthetics per side, such as color, width etc... The package also includes a vignette covering it's main usecases - symmetric, asymmetric, one-sided, and per-side styling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iamYannC/ggerror}
  \item \url{https://iamyannc.github.io/ggerror/}
  \item Report bugs at \url{https://github.com/iamYannC/ggerror/issues}
}

}
\author{
\strong{Maintainer}: Yann Cohen \email{yannco5@gmail.com} (\href{https://orcid.org/0009-0009-0509-3609}{ORCID})

}
\keyword{internal}
